xform: move groupby.opt functions to groupby_funcs.go
Release note: None
